User by id middleware
